Uptime: fix for uptime calculation error which resulted in very large number in "channel state info" Version increment 23.07

+++ extras/porting/linux/supla/control/cmd_relay.h
@@ -45,7 +45,7 @@ class CmdRelay : public Sensor::SensorParsed<VirtualRelay> {
  protected:
   std::string cmdOn;
   std::string cmdOff;
-  uint64_t lastReadTime = 0;
+  uint32_t lastReadTime = 0;
   bool useOfflineOnInvalidState = false;
 };
