+++ extras/test/SuplaDeviceTests/supla_device_tests.cpp
@@ -45,8 +45,8 @@ class SuplaDeviceTests : public ::testing::Test {
 
 class TimeInterfaceStub : public TimeInterface {
   public:
-    virtual uint64_t millis() override {
-      static uint64_t value = 0;
+    virtual uint32_t millis() override {
+      static uint32_t value = 0;
       value += 1000;
       return value;
     }
