+++ src/SuplaDevice.cpp
@@ -385,7 +385,7 @@ void SuplaDeviceClass::iterate(void) {
     cfg->saveIfNeeded();
   }
 
-  uint64_t _millis = millis();
+  uint32_t _millis = millis();
   checkIfRestartIsNeeded(_millis);
   handleLocalActionTriggers();
   iterateAlwaysElements(_millis);
@@ -640,7 +640,7 @@ bool SuplaDeviceClass::loadDeviceConfig() {
   return configComplete;
 }
 
-void SuplaDeviceClass::iterateAlwaysElements(uint64_t _millis) {
+void SuplaDeviceClass::iterateAlwaysElements(uint32_t _millis) {
   uptime.iterate(_millis);
 
   // Iterate all elements
@@ -1015,7 +1015,7 @@ void SuplaDeviceClass::handleLocalActionTriggers() {
   }
 }
 
-void SuplaDeviceClass::checkIfRestartIsNeeded(uint64_t _millis) {
+void SuplaDeviceClass::checkIfRestartIsNeeded(uint32_t _millis) {
   if (deviceRestartTimeoutTimestamp != 0 &&
       _millis - deviceRestartTimeoutTimestamp > 5ul * 60 * 1000) {
     SUPLA_LOG_INFO("Config mode 5 min timeout. Reset device");
