+++ src/supla/sensor/thermometer.h
@@ -34,7 +34,7 @@ class Thermometer : public ChannelElement {
   void iterateAlways() override;
 
  protected:
-  uint64_t lastReadTime;
+  uint32_t lastReadTime;
 };
 
 };  // namespace Sensor
